requirements.txt
peewee==3.13.3
PyMySQL
mysql-connector-python